+++ app/layout.tsx
@@ -4,6 +4,7 @@ import { geistMono, geistSans } from "@/app/ui/fonts";
 import "@/app/globals.css";
 import TopNav from "@/app/ui/navigation/topnav";
 import { ThemeProvider } from "@/app/ui/navigation/theme-provider";
+import Footer from "@/app/ui/footer/footer";
 
 export const metadata: Metadata = {
   title: "PROMPTZ - Discover, Create, and Share Prompts for Amazon Q Developer",
@@ -70,6 +71,7 @@ export default function RootLayout({
           <div className="min-h-screen bg-gradient-to-b from-black via-purple-950/20 to-black overflow-hidden">
             <div className="max-w-7xl mx-auto ">{children}</div>
           </div>
+          <Footer />
         </ThemeProvider>
         {/* <AuthProvider>
           <TopNav />
